+++ lib/path/component/collection.dart
@@ -142,7 +142,7 @@ abstract class Collection<TChild extends IChild>
   }
 
   /// Get the protocol of the path.
-  String get protocol => Protocol.tmp;
+  String get protocol;
 
   /// Get the path.
   String get path => this._path.url;