+++ lib/path/component/unit.dart
@@ -130,7 +130,7 @@ abstract class Unit<TValue extends Object>
   }
 
   /// Get the protocol of the path.
-  String get protocol => Protocol.tmp;
+  String get protocol;
 
   /// Get the path.
   String get path => this._path.url;